move controls buttons to header_box
